chore: write first version of parser functions
